Make support/rspec.rb closer to RSpec default
Having this file be completely devoid of comments makes it harder for people to understand, particularly when many of the RSpec configuration options have significant impact on the features available when using the tool.

`rspec —init`, on the other hand, creates a spec_helper.rb with *lots* of comments. Probably a little too much.

As a compromise, copy the most helpful aspects of the comments from the default spec_helper.rb to our support/rspec.rb, condensing the text wherever possible.

In addition, enable the one feature from the `rspec —init` output that was missing in our file: `config.example_status_persistence_file_path`. This is a very helpful feature and it would be good for our users to have access to this by default. Leave this configured with the default “spec/examples.txt”, and in our `hanami install` hook, update the `.gitignore` file to ignore it.