Skip generating tests with --skip-tests flag given
Having the ability to skip tests when generating an action or part is useful if you know you're going to test it in _some other way_ than what we generate by default.

For example, when working on a web app, and when using Capybara-driven feature specs, we generally don't want request specs generated for each action class.