Disable warnings in RSpec by default
Even when running tests for the getting started guide, this was resulting in some warnings being output (for e.g. the debug gem). To reduce noisy and give people greater confidence with their first tests, disable this by default.

Leave a comment recommending that the user turn it on if they choose.