Make support/rspec.rb closer to RSpec default (#20)
Having this file be completely devoid of comments makes it harder for people to understand, particularly when many of the RSpec configuration options have significant impact on the features available when using the tool.

`rspec --init`, on the other hand, creates a spec_helper.rb with *lots* of comments. Probably a little too much.

As a compromise, copy the most helpful aspects of the comments from the default spec_helper.rb to our support/rspec.rb, condensing the text wherever possible.

In addition, enable the one feature from the `rspec —init` output that was missing in our file: `config.example_status_persistence_file_path`. This is a very helpful feature and it would be good for our users to have access to this by default. Leave this configured with the default “spec/examples.txt”, and in our `hanami install` hook, update the `.gitignore` file to ignore it.

In addition, make a couple of other adjustments:

- Comment out the `config.warnings = true` RSpec configuration by default: even when running tests for the getting started guide, this was resulting in some warnings being output (for e.g. the debug gem). To reduce noise and give people greater confidence with their first tests, disable this by default. Leave a comment recommending that the user turn it on if they choose.
- Use a clearer name for request spec shared context: "Hanami app" was too generic. By using "Rack::Test" here, we make it clearer that this shared context is specific for Rack::Test specs only.